cleanup: Move query out of database header
