+++ include/osquery/distributed.h
@@ -13,7 +13,7 @@
 #include <string>
 #include <vector>
 
-#include <osquery/database.h>
+#include <osquery/query.h>
 #include <osquery/registry.h>
 #include <osquery/status.h>
