+++ include/osquery/sql.h
@@ -14,8 +14,8 @@
 #include <string>
 #include <vector>
 
-#include <osquery/database.h>
 #include <osquery/flags.h>
+#include <osquery/query.h>
 #include <osquery/registry.h>
 #include <osquery/tables.h>
