+++ osquery/CMakeLists.txt
@@ -36,6 +36,7 @@ if(POSIX)
     -Wno-zero-length-array
     -Wno-unused-parameter
     -Wno-gnu-case-range
+    -Weffc++
   )
   if(NOT FREEBSD)
     add_compile_options(
