+++ osquery/carver/carver.cpp
@@ -16,6 +16,7 @@
 
 #include <boost/property_tree/ptree.hpp>
 
+#include <osquery/database.h>
 #include <osquery/distributed.h>
 #include <osquery/flags.h>
 #include <osquery/logger.h>
@@ -25,9 +26,7 @@
 #include "osquery/core/conversions.h"
 #include "osquery/core/json.h"
 #include "osquery/filesystem/fileops.h"
-#include "osquery/remote/requests.h"
 #include "osquery/remote/serializers/json.h"
-#include "osquery/remote/transports/tls.h"
 #include "osquery/remote/utility.h"
 #include "osquery/tables/system/hash.h"
