+++ osquery/config/packs.cpp
@@ -12,6 +12,7 @@
 #include <random>
 
 #include <osquery/core.h>
+#include <osquery/database.h>
 #include <osquery/logger.h>
 #include <osquery/packs.h>
 #include <osquery/sql.h>
