+++ osquery/core/CMakeLists.txt
@@ -1,3 +1,10 @@
+#  Copyright (c) 2014-present, Facebook, Inc.
+#  All rights reserved.
+#
+#  This source code is licensed under the BSD-style license found in the
+#  LICENSE file in the root directory of this source tree. An additional grant
+#  of patent rights can be found in the PATENTS file in the same directory.
+
 file(GLOB OS_CORE_SOURCE "${PROCESS_FAMILY}/*.cpp")
 file(GLOB OS_CORE_TESTS_SOURCE "tests/${PROCESS_FAMILY}/*.cpp")
 
@@ -16,6 +23,7 @@ ADD_OSQUERY_LIBRARY(TRUE osquery_core
   tables.cpp
   flags.cpp
   watcher.cpp
+  query.cpp
 )
 
 file(GLOB OSQUERY_CORE_TESTS "tests/*.cpp")
