+++ osquery/events/CMakeLists.txt
@@ -1,3 +1,10 @@
+#  Copyright (c) 2014-present, Facebook, Inc.
+#  All rights reserved.
+#
+#  This source code is licensed under the BSD-style license found in the
+#  LICENSE file in the root directory of this source tree. An additional grant
+#  of patent rights can be found in the PATENTS file in the same directory.
+
 if(APPLE)
   ADD_OSQUERY_LINK_ADDITIONAL("-framework CoreServices")
   ADD_OSQUERY_LINK_ADDITIONAL("-framework SystemConfiguration")
